+++ src/main/groovy/life/qbic/registration/registries/NfCoreResultRegistry.groovy
@@ -57,7 +57,7 @@ class NfCoreResultRegistry implements Registry {
         tmpMap.put("nf-core/chipseq", QExperimentType.Q_WF_NGS_CHIP_SEQ)
         tmpMap.put("nf-core/eager", QExperimentType.Q_WF_NGS_ANCIENT_DNA_SEQ)
         tmpMap.put("nf-core/hlatyping", QExperimentType.Q_WF_NGS_HLATYPING)
-        tmpMap.put("nf-core/mag", QExperimentType.Q_WF_NGS_SHOTGUN_METAGENOMICS_ANALYIS)
+        tmpMap.put("nf-core/mag", QExperimentType.Q_WF_NGS_SHOTGUN_METAGENOMICS)
         tmpMap.put("nf-core/methylseq", QExperimentType.Q_WF_NGS_METHYLATION_ANALYSIS)
         tmpMap.put("nf-core/methaboigniter", QExperimentType.Q_WF_MX_IDENTIFY_AND_QUANTIFY)
         tmpMap.put("nf-core/mhcquant", QExperimentType.Q_WF_LX_MHC_QUANTIFICATION)
@@ -171,23 +171,6 @@ class NfCoreResultRegistry implements Registry {
         }
     }
 
-    private static Optional<Context> getContext(SampleId sampleId,
-                                                ISearchService searchService) {
-        SearchCriteria sc = new SearchCriteria()
-        sc.addMatchClause(
-                SearchCriteria.MatchClause.createAttributeMatch(SearchCriteria.MatchClauseAttribute.CODE, sampleId.toString())
-        )
-        List<ISampleImmutable> searchResult = searchService.searchForSamples(sc)
-        if (!searchResult) {
-            return Optional.empty()
-        }
-        ProjectSpace space = new ProjectSpace(searchResult[0].getSpace())
-        ProjectCode code = sampleId.getProjectCode()
-
-        Context context = new Context(projectSpace: space, projectCode: code)
-        return Optional.of(context)
-    }
-
     /*
     Does the final registration of the dataset in openBIS.
      */
@@ -197,10 +180,10 @@ class NfCoreResultRegistry implements Registry {
         // 1. Get the openBIS samples the datasets belong to
         // Will contain the openBIS samples which data served as input data for
         // the pipeline run
-        List<SampleId> sampleIdList = validateSampleIds(sampleIds)
+        List<SampleId> sampleIdList = Utils.validateSampleIds(sampleIds)
 
-        this.context = getContext(sampleIdList[0], transaction.getSearchService()).orElseThrow({
-            new RegistrationException("Could not determine context for samples ${sampleIdList}")
+        this.context = Utils.getContext(sampleIdList[0], transaction.getSearchService()).orElseThrow({
+            throw new RegistrationException("Could not determine context for samples ${sampleIdList}")
         })
 
         List<ISample> parentSamples = []
@@ -276,16 +259,6 @@ class NfCoreResultRegistry implements Registry {
 
         // 9. Attach result data to dataset
         transaction.moveFile(this.datasetRootPath.toString(), dataset)
-    }
-
-    private List<SampleId> validateSampleIds(List<String> sampleIdList) throws RuntimeException {
-        def convertedSampleIds = []
-        for(String sampleId : sampleIdList) {
-            def convertedId = SampleId.from(sampleId).orElseThrow( {
-                throw new RuntimeException("$sampleId does not seem to contain a valid sample id.")})
-            convertedSampleIds.add(convertedId)
-        }
-        return convertedSampleIds
     }
 
     /*
