+++ src/main/groovy/life/qbic/registration/types/QDatasetType.groovy
@@ -8,6 +8,7 @@ package life.qbic.registration.types
 enum QDatasetType {
     Q_WF_LX_MHC_QUANTIFICATION_RESULTS,
     Q_WF_MX_IDENTIFY_AND_QUANTIFY_RESULTS,
+    Q_WF_MS_MAXQUANT_RESULTS,
     Q_WF_NGS_16S_AMPLICON_SEQ_RESULTS,
     Q_WF_NGS_ATAC_SEQ_RESULTS,
     Q_WF_NGS_ANCIENT_DNA_SEQ_RESULTS,
